Update examples for new channel API and rename Channel to AudioChannel
