+++ examples/README.md
@@ -1,8 +1,17 @@
-# Example for Kira audio in Bevy
+# Examples
 
-This is a simple Bevy App illustrating the capabilities of `bevy_kira_audio`. Run the example with `cargo run --example basic --features example`.
+These examples are simple Bevy Apps illustrating the capabilities of `bevy_kira_audio`. Run the examples with `cargo run --example <example>`.
+
+| Example                                         | Description                                                          |
+|-------------------------------------------------|----------------------------------------------------------------------|
+| [`basic.rs`](/basic.rs)                         | Display of basic functionality                                       |
+| [`custom_channel.rs`](/custom_channel.rs)       | How to add and use a custom audio channel                            |
+| [`multiple_channels.rs`](/multiple_channels.rs) | GUI application with full control over tree different audio channels |
+| [`semantic_duration.rs`](/semantic_duration.rs) | Load an audio file with custom semantic duration from a ron asset    |
+| [`status.rs`](/status.rs)                       | Continuously get the playback state of a sound                       |
+| [`stream.rs`](/stream.rs)                       | Play an audio stream                                                 |
 
 ## Credits
-The example includes copyright assets:
+The examples include assets with copyright:
 
 Loop audio: [CC BY 3.0](https://creativecommons.org/licenses/by/3.0/) [Jay_You](https://freesound.org/people/Jay_You/sounds/460432/)
