+++ src/lib.rs
@@ -4,7 +4,7 @@
 //! via Bevy's ECS.
 //!
 //! ```edition2018
-//! # use bevy_kira_audio::{AudioChannel, Audio, AudioPlugin};
+//! # use bevy_kira_audio::{AudioStreamChannel, Audio, AudioPlugin};
 //! # use bevy::prelude::*;
 //! # use bevy::asset::AssetPlugin;
 //! # use bevy::app::AppExit;
@@ -31,8 +31,8 @@
 #![forbid(unsafe_code)]
 #![warn(unused_imports, missing_docs)]
 
-pub use audio::{AudioApp, Channel, InstanceHandle, PlaybackState};
-pub use channel::AudioChannel;
+pub use audio::{AudioApp, AudioChannel, InstanceHandle, PlaybackState};
+pub use channel::AudioStreamChannel;
 pub use source::AudioSource;
 pub use stream::{AudioStream, Frame, StreamedAudio};
 
@@ -72,7 +72,7 @@ compile_error!("You need to enable at least one of the bevy_kira_audio features 
 /// Add this plugin to your Bevy app to get access to
 /// the Audio resource
 /// ```edition2018
-/// # use bevy_kira_audio::{AudioChannel, Audio, AudioPlugin};
+/// # use bevy_kira_audio::{AudioStreamChannel, Audio, AudioPlugin};
 /// # use bevy::prelude::*;
 /// # use bevy::asset::AssetPlugin;
 /// # use bevy::app::AppExit;
@@ -118,7 +118,7 @@ impl Plugin for AudioPlugin {
             CoreStage::PreUpdate,
             cleanup_stopped_instances.label(AudioSystemLabel::InstanceCleanup),
         )
-        .add_channel::<Audio>();
+        .add_audio_channel::<MainTrack>();
     }
 }
 
@@ -129,15 +129,23 @@ pub(crate) enum AudioSystemLabel {
 
 /// The default audio channel
 ///
-/// Use it as a [`Channel<Audio>`] resource in your systems.
-/// You can add your own channels via [`add_channel`](audio::AudioApp::add_channel).
-pub struct Audio;
+/// Alias for the [`AudioChannel<MainTrack>`] resource. Use it to play and control sound on the main track.
+/// You can add your own channels via [`add_audio_channel`](audio::AudioApp::add_audio_channel).
+pub type Audio = AudioChannel<MainTrack>;
+
+/// Type for the default audio channel
+///
+/// Use it via the [`AudioChannel<MainTrack>`] resource to play and control sound on the main track.
+/// You can add your own channels via [`add_audio_channel`](audio::AudioApp::add_audio_channel).
+///
+/// You can use [`Audio`] as a type alias for [`AudioChannel<MainTrack>`]
+pub struct MainTrack;
 
 /// A Bevy plugin for streaming of audio
 ///
 /// This plugin requires [AudioPlugin] to also be active
 /// ```edition2018
-/// # use bevy_kira_audio::{AudioStream, Frame, StreamedAudio, AudioChannel, Audio, AudioPlugin, AudioStreamPlugin};
+/// # use bevy_kira_audio::{AudioStream, Frame, StreamedAudio, AudioStreamChannel, Audio, AudioPlugin, AudioStreamPlugin};
 /// # use bevy::prelude::*;
 /// # use bevy::asset::AssetPlugin;
 /// # use bevy::app::AppExit;
